Move implementation of metrics.FileWriter to metrics/file_writer.go
